Partially added world. Also singleton and logger

+++ Gesco/src/World.h
@@ -10,9 +10,9 @@
 
 #include <opencv2/opencv.hpp>
 #include <opencv2/highgui/highgui.hpp>
-#include <SDL.h>
+//#include <SDL.h>
 
-#include "VideoFactory.h"
+//#include "VideoFactory.h"
 #include "Logger.h"
 #include "Singleton.h"
 
@@ -22,7 +22,7 @@
 class World: public Singleton<World> {
 	int _width;
 	int _height;
-	Uint32 _ticksLastEvent;
+	uint64 _ticksLastEvent;
 
 
 public:
